backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
google-cloud-bigquery==3.25.0
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.2
python-multipart==0.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.44.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
matplotlib==3.8.4
